pyproject.toml
[tool.black]
line-length = 80
target-version = ['py39']
exclude = '/.*(__pycache__).*/'

[tool.setuptools]
py-modules = ["srt_translator"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "srt_translator"
version = "1.0.0"
description = "A python translator for srt files."
authors = [
    {name="Connor Makowski", email="conmak@mit.edu"}
]
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    'License :: OSI Approved :: MIT License',
    "Operating System :: OS Independent",
]
dependencies = [
    "google-cloud-translate==3.12.1", 
    "type_enforced>=1.2.0",
]

[project.urls]
"Homepage" = "https://github.com/connor-makowski/srt_translator"
"Bug Tracker" = "https://github.com/connor-makowski/srt_translator/issues"
"Documentation" = "https://connor-makowski.github.io/srt_translator/srt_translator/enforcer.html"